rewrite tests, revert to default format on inconsistent request
