+++ bin/orca.js
@@ -2,7 +2,7 @@
 
 const path = require('path')
 const { spawn } = require('child_process')
-const electronPath = require('electron')
+const electronPath = process.env['ELECTRON_PATH'] ? process.env['ELECTRON_PATH'] : require('electron')
 
 const args = process.argv.slice(2)
 const pathToMain = path.join(__dirname, 'orca_electron.js')
